+++ Rdmp.Core/DataViewing/ViewAggregateExtractUICollection.cs
@@ -6,11 +6,13 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using FAnsi.Discovery.QuerySyntax;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.Curation.Data.Aggregation;
 using Rdmp.Core.Curation.Data.Dashboarding;
+using Rdmp.Core.DataExport.DataExtraction;
 using Rdmp.Core.Providers;
 using Rdmp.Core.QueryBuilding;
 using ReusableLibraryCode.DataAccess;
@@ -58,7 +60,6 @@ namespace Rdmp.Core.DataViewing
 
             return null;
         }
-
 
         public IDataAccessPoint GetDataAccessPoint()
         {